Reuse framework transform function
Function was abstracted out recently so there
is now a duplicate. This change removes the
duplicate and moves the tests.
