+++ src/apps/transformers.js
@@ -3,14 +3,9 @@ const {
   filter,
   keyBy,
   snakeCase,
-  isPlainObject,
-  isFunction,
-  isEmpty,
-  assign,
 } = require('lodash')
 const { isValid, format, parse } = require('date-fns')
 
-const { buildPagination } = require('../lib/pagination')
 const { hqLabels } = require('./companies/labels')
 
 function transformObjectToOption ({ id, name }) {
@@ -94,40 +89,6 @@ function buildMetaDataObj (collection) {
   })
 }
 
-/**
- * @param {object} [options] {object}
- * @param {string} [options.searchTerm] - search term used for highlighting words in collection macro
- * @param {object} [options.query] - URL query object used in pagination
- * @param {...function} [itemTransformers] - an array of transformer functions to apply for each item in the list
- * @returns {object, undefined}
- */
-function transformApiResponseToCollection (options = {}, ...itemTransformers) {
-  /**
-   * @param {object} response - API response object
-   * @returns {function}
-   */
-  return function transformResponseToCollection (response) {
-    if (!isPlainObject(response)) { return }
-
-    let items = response.results || response.items
-
-    if (!items) { return }
-
-    itemTransformers.forEach(transformer => {
-      if (!isFunction(transformer)) { return }
-      items = items
-        .map(transformer)
-        .filter(item => !isEmpty(item))
-    })
-
-    return assign({}, {
-      items,
-      count: response.count,
-      pagination: buildPagination(options.query, response),
-    })
-  }
-}
-
 module.exports = {
   buildMetaDataObj,
   transformHQCodeToLabelledOption,
@@ -135,7 +96,6 @@ module.exports = {
   transformStringToOption,
   transformContactToOption,
   transformIdToObject,
-  transformApiResponseToCollection,
   transformDateObjectToDateString,
   transformDateStringToDateObject,
 }
